+++ Changelog.md
@@ -7,6 +7,8 @@
 * Python overlay
     * Add a hide panel button.
     * Timer hotkeys not available when hidden window.
+    * AoE2
+        * Bengalis time evalutation affected by the 2 villagers bonus.
 
 ## [1.9.1] - 2024.03.17
 * Option to keep the BO writer and hotkeys configuration always on top.